requirements-dev.txt
beautifulsoup4
black
django-debug-toolbar
django-extensions
freezegun
pydocstyle
pylint
pylint-django
pytest
pytest-cov
pytest-django
pytest-randomly
pytest-sugar
